Unify tests between -sys and -js API
